Clean code: move all date, time and duration utils into the same package 'org.opentripplanner.util.time'.
