+++ src/main/java/org/opentripplanner/transit/raptor/api/path/PathLeg.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.transit.raptor.api.path;
 
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
-import org.opentripplanner.transit.raptor.util.TimeUtils;
+import org.opentripplanner.util.time.TimeUtils;
 
 /**
  * A leg in a Raptor path. The legs are linked together from the first leg {@link AccessPathLeg},
