+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/standard/stoparrivals/path/EgressArrivalToPathAdapter.java
@@ -10,7 +10,7 @@ import org.opentripplanner.transit.raptor.rangeraptor.standard.stoparrivals.Egre
 import org.opentripplanner.transit.raptor.rangeraptor.standard.stoparrivals.view.StopsCursor;
 import org.opentripplanner.transit.raptor.rangeraptor.transit.TransitCalculator;
 import org.opentripplanner.transit.raptor.rangeraptor.view.DebugHandler;
-import org.opentripplanner.transit.raptor.util.TimeUtils;
+import org.opentripplanner.util.time.TimeUtils;
 
 
 /**
