+++ src/main/java/org/opentripplanner/util/time/TimeUtils.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.transit.raptor.util;
+package org.opentripplanner.util.time;
 
 
 import javax.annotation.Nonnull;
