+++ src/test/java/org/opentripplanner/transit/raptor/speed_test/testcase/Result.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.transit.raptor.speed_test.testcase;
 
 import org.opentripplanner.routing.core.TraverseMode;
-import org.opentripplanner.transit.raptor.util.TimeUtils;
+import org.opentripplanner.util.time.TimeUtils;
 
 import java.util.ArrayList;
 import java.util.Comparator;
