+++ src/test/java/org/opentripplanner/transit/raptor/speed_test/testcase/TestCase.java
@@ -2,7 +2,7 @@ package org.opentripplanner.transit.raptor.speed_test.testcase;
 
 import org.opentripplanner.transit.raptor.speed_test.model.Itinerary;
 import org.opentripplanner.transit.raptor.speed_test.model.Place;
-import org.opentripplanner.transit.raptor.util.TimeUtils;
+import org.opentripplanner.util.time.TimeUtils;
 
 import java.util.Collection;
 import java.util.List;
